move logup in its own file to be easier to import somewhere else
